DesignMenu: move all tools under design-tools
