+++ client/my-sites/design-menu/design-tools/header-image-control.js
