+++ client/my-sites/design-menu/design-tools/site-title-control.js
